+++ src/commands/mod.rs
@@ -1,4 +1,3 @@
-pub mod can_can;
 pub mod can_monitor;
 pub mod default_config;
 pub mod gateway;